requirements.txt
pathlib~=1.0.1
rich~=14.0.0
python-dotenv~=1.1.0
PyYAML~=6.0.2
Jinja2~=3.1.6
torch~=2.7.0
filelock~=3.18.0
chromadb>=1.0.0
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-chrome>=0.2.0
langchain-openai>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langchain-huggingface>=0.1.0
openai>=1.75.0
anthropic>=0.50.0
huggingface-hub>=0.30.0
pyyaml>=6.0.2
python-dotenv>=1.1.0
fastapi>=0.115.0
uvicorn>=0.34.0
rich>=14.0.0
jsonschema>=4.23.0
pydantic>=2.11.0
pytest>=8.3.5
pytest-asyncio>=0.26.0
httpx>=0.28.1
tenacity>=9.1.2
tiktoken>=0.9.0
numpy>=2.2.5
tqdm>=4.67.1
transformers>=4.51.3